AbstractLogSercive: extract class LogLevelStrategy
Move code for log level estimation to new class LogLevelStrategy
to separate responsibilities.
