+++ src/main/java/org/scijava/log/AbstractLogService.java
@@ -31,10 +31,6 @@
 
 package org.scijava.log;
 
-import java.util.HashMap;
-import java.util.Map;
-import java.util.Properties;
-
 import org.scijava.service.AbstractService;
 
 /**
@@ -47,85 +43,27 @@ public abstract class AbstractLogService extends AbstractService implements
 	LogService
 {
 
-	private int currentLevel = levelFromEnvironment();
-
-	private final Map<String, Integer> classAndPackageLevels =
-		new HashMap<>();
-
-	// -- constructor --
-
-	public AbstractLogService() {
-		// check SciJava log level system properties for initial logging levels
-
-		// global log level property
-		final String logProp = System.getProperty(LOG_LEVEL_PROPERTY);
-		final int level = LogLevel.value(logProp);
-		if (level >= 0) setLevel(level);
-
-		if (getLevel() == 0)
-			setLevel(levelFromEnvironment());
-
-		// populate custom class- and package-specific log level properties
-		final String logLevelPrefix = LOG_LEVEL_PROPERTY + ":";
-		final Properties props = System.getProperties();
-		for (final Object propKey : props.keySet()) {
-			if (!(propKey instanceof String)) continue;
-			final String propName = (String) propKey;
-			if (!propName.startsWith(logLevelPrefix)) continue;
-			final String classOrPackageName =
-				propName.substring(logLevelPrefix.length());
-			setLevel(classOrPackageName, LogLevel.value(props.getProperty(propName)));
-		}
-	}
+	private LogLevelStrategy logLevelStrategy = new LogLevelStrategy();
 
 	// -- Logger methods --
 
 	@Override
 	public int getLevel() {
-		if (!classAndPackageLevels.isEmpty()) {
-			// check for a custom log level for calling class or its parent packages
-			String classOrPackageName = callingClass();
-			while (classOrPackageName != null) {
-				final Integer level = classAndPackageLevels.get(classOrPackageName);
-				if (level != null) return level;
-				classOrPackageName = parentPackage(classOrPackageName);
-			}
-		}
-		// no custom log level; return the global log level
-		return currentLevel;
+		return logLevelStrategy.getLevel();
 	}
 
 	@Override
 	public void setLevel(final int level) {
-		currentLevel = level;
+		logLevelStrategy.setLevel(level);
 	}
 
 	@Override
 	public void setLevel(final String classOrPackageName, final int level) {
-		classAndPackageLevels.put(classOrPackageName, level);
+		logLevelStrategy.setLevel(classOrPackageName, level);
 	}
 
-	// -- Helper methods --
-
-	private String callingClass() {
-		final String thisClass = AbstractLogService.class.getName();
-		for (final StackTraceElement element : new Exception().getStackTrace()) {
-			final String className = element.getClassName();
-			// NB: Skip stack trace elements from other methods of this class.
-			if (!thisClass.equals(className)) return className;
-		}
-		return null;
-	}
-
-	private String parentPackage(final String classOrPackageName) {
-		final int dot = classOrPackageName.lastIndexOf(".");
-		if (dot < 0) return null;
-		return classOrPackageName.substring(0, dot);
-	}
-
-	private int levelFromEnvironment() {
-		// use the default, which is INFO unless the DEBUG env. variable is set
-		return System.getenv("DEBUG") == null ? LogLevel.INFO : LogLevel.DEBUG;
-	}
+	@Override
+	public abstract void alwaysLog(final int level, final Object msg,
+		final Throwable t);
 
 }
